+++ config.py
@@ -16,6 +16,7 @@ class Settings(BaseSettings):
 
     google_gemini_api_key: str | None = None
     anthropic_api_key: str | None = None
+    telegra_ph_token: str | None = None
 
     @cached_property
     def openai_client(self) -> openai.OpenAI:
@@ -23,6 +24,12 @@ class Settings(BaseSettings):
             api_key=self.openai_api_key,
             base_url=self.openai_base_url,
         )
+
+    @cached_property
+    def telegraph_client(self):
+        from handlers._telegraph import TelegraphAPI
+
+        return TelegraphAPI(self.telegra_ph_token)
 
 
 settings = Settings()  # type: ignore
