+++ handlers/claude.py
@@ -1,16 +1,15 @@
+import time
 from os import environ
 from pathlib import Path
-import time
 
 from anthropic import Anthropic, APITimeoutError
+from expiringdict import ExpiringDict
 from telebot import TeleBot
 from telebot.types import Message
-from expiringdict import ExpiringDict
-
-from . import *
-
 from telegramify_markdown import convert
 from telegramify_markdown.customize import markdown_symbol
+
+from ._utils import bot_reply_first, bot_reply_markdown, enrich_text_with_urls
 
 markdown_symbol.head_level_1 = "📌"  # If you want, Customizing the head level 1 symbol
 markdown_symbol.link = "🔗"  # If you want, Customizing the link symbol
